# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plata-Forma API
  version: 0.0.1
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid data or user exists
  /auth/login:
    post:
      summary: Login and get a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid data
        '401':
          description: Invalid credentials
        '429':
          description: Account locked
  /households/{householdId}/transactions:
    post:
      summary: Create a transaction
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
  /households/{householdId}/reports/monthly:
    get:
      summary: Monthly report
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
        - name: month
          in: query
          required: true
          schema:
            type: string
            example: 2026-02
      responses:
        '200':
          description: Monthly report summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyReportResponse'
  /transactions/{id}:
    patch:
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionRequest'
      responses:
        '200':
          description: Transaction updated
    delete:
      summary: Delete (soft) a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction archived
  /households/{householdId}/transactions/balances:
    get:
      summary: Monthly balances by account and category
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
        - name: month
          in: query
          required: true
          schema:
            type: string
            example: 2026-02
      responses:
        '200':
          description: Monthly balances
  /households/{householdId}/categories/balances:
    get:
      summary: Monthly and yearly category totals
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
        - name: month
          in: query
          required: true
          schema:
            type: string
            example: 2026-02
      responses:
        '200':
          description: Category totals
  /households/{householdId}/accounts/balances:
    get:
      summary: Account totals
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account totals
  /credit-cards/{id}:
    delete:
      summary: Delete (soft) a credit card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit card archived
  /loans/{id}:
    delete:
      summary: Delete (soft) a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Loan archived
components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    RegisterResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        userId:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        accessToken:
          type: string
        tokenType:
          type: string
        userId:
          type: string
    TransactionEntry:
      type: object
      required: [from, to, amount]
      properties:
        from:
          $ref: '#/components/schemas/EntrySide'
        to:
          $ref: '#/components/schemas/EntrySide'
        amount:
          type: string
          description: Amount in cents as an integer string
        memo:
          type: string
    EntrySide:
      type: object
      required: [kind, id]
      properties:
        kind:
          type: string
          enum: [ACCOUNT, CATEGORY]
        id:
          type: string
    TransactionLine:
      type: object
      required: [accountId, amount]
      properties:
        accountId:
          type: string
        categoryId:
          type: string
        amount:
          type: string
        memo:
          type: string
    CreateTransactionRequest:
      type: object
      required: [date]
      properties:
        date:
          type: string
          example: 2026-02-19
        description:
          type: string
        entry:
          $ref: '#/components/schemas/TransactionEntry'
        lines:
          type: array
          items:
            $ref: '#/components/schemas/TransactionLine'
    UpdateTransactionRequest:
      type: object
      properties:
        date:
          type: string
          example: 2026-02-19
        description:
          type: string
        entry:
          $ref: '#/components/schemas/TransactionEntry'
        lines:
          type: array
          items:
            $ref: '#/components/schemas/TransactionLine'
    MonthlyReportResponse:
      type: object
      properties:
        month:
          type: string
        range:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        totals:
          type: object
          properties:
            income:
              type: string
            expense:
              type: string
            net:
              type: string
        byCategory:
          type: array
          items:
            type: object
            properties:
              categoryId:
                type: string
              name:
                type: string
              amount:
                type: string
        byAccount:
          type: array
          items:
            type: object
            properties:
              accountId:
                type: string
              name:
                type: string
              type:
                type: string
              amount:
                type: string